Abstracted methods for TestingService
