+++ src/DevelopmentHell.Hubba/DevelopmentHell.Hubba.Authorization.Test/DevelopmentHell.Hubba.Authorization.Test.csproj
@@ -22,6 +22,7 @@
   <ItemGroup>
     <ProjectReference Include="..\DevelopmentHell.Hubba.Authorization.Service\DevelopmentHell.Hubba.Authorization.Service.csproj" />
     <ProjectReference Include="..\DevelopmentHell.Hubba.Registration\DevelopmentHell.Hubba.Registration.Service.csproj" />
+    <ProjectReference Include="..\DevelopmentHell.Hubba.Testing.Service\DevelopmentHell.Hubba.Testing.Service.csproj" />
   </ItemGroup>
 
 </Project>
